Rename files and fix linking. Add Manage Notification Links to User Header Dropdown menu as well as Community SideBar.
