+++ client/scripts/views/components/header.ts
@@ -105,6 +105,13 @@ const Header: m.Component<{}> = {
             iconLeft: Icons.SETTINGS,
             label: 'Settings'
           }),
+          (app.vm.activeAccount
+            || m.route.get() === '/settings')
+            && m(MenuItem, {
+              onclick: () => m.route.set('/notifications'),
+              iconLeft: Icons.BELL,
+              label: 'Manage Notifications'
+            }),
           app.login?.isSiteAdmin && app.activeChainId() && m(MenuItem, {
             onclick: () => m.route.set(`/${app.activeChainId()}/admin`),
             iconLeft: Icons.USER,
